Add Semigroup instances for Monoids
This is based on the advice from
https://prime.haskell.org/wiki/Libraries/Proposals/SemigroupMonoid#Writingcompatiblecode

The changes to the postgresql backend are more extensive, and
less pleasing, due to the need to work around the lack of Semigroup
support in the postgresql-simple types for old versions of the package.
There may well be a better way to achieve this.